dendrogram_data {ggdendro}	R Documentation

Extract data frame from dendrogram object for plotting using ggplot.

Description

Extract data frame from dendrogram object for plotting using ggplot

Usage

dendrogram_data(x, type = c("rectangle", "triangle"), ...)

Arguments

x
object of class "dendrogram", e.g. the output of as.dendrogram()

type
The type of plot, indicating the shape of the dendrogram: "rectangle" will draw rectangular lines, while "triangle" will draw triangular lines.

...
ignored

See Also

ggdendrogram()

Other dendro_data methods: dendro_data(), dendro_data.rpart(), dendro_data.tree(), rpart_labels()

Other dendrogram/hclust functions: dendro_data()

[Package ggdendro version 0.2.0 Index]