rpart_segments {ggdendro}	R Documentation

Extract data frame from rpart object for plotting using ggplot.

Description

Extract data frame from rpart object for plotting using ggplot.

Usage

rpart_segments(x, ...)

Arguments

...
ignored

See Also

ggdendrogram()

Other rpart functions: dendro_data.rpart(), rpart_labels()

[Package ggdendro version 0.2.0 Index]