tree_segments {ggdendro}	R Documentation

Extract data frame from tree object for plotting using ggplot.

Description

Extract data frame from tree object for plotting using ggplot.

Usage

tree_segments(model, uniform, ...)

Arguments

model
object of class "tree", e.g. the output of tree()

...
ignored

Author(s)

Code modified from original by Brian Ripley

See Also

ggdendrogram()

Other tree functions: dendro_data.tree(), get_data_tree_leaf_labels(), tree_labels()

[Package ggdendro version 0.2.0 Index]